% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_normal_normal_data}
\alias{generate_normal_normal_data}
\title{Generate Normal-Normal Data}
\usage{
generate_normal_normal_data(num_groups, sample_sizes = NULL,
  min_sample_size = 1, max_sample_size = 100, hyper_mean = 0,
  hyper_var = 1, data_var = 1, seed = NULL)
}
\arguments{
\item{num_groups}{Number of groups or clusters}
}
\description{
Function to generate data according to normal-normal heirarchical model
}
\examples{
generate_normal_normal_data()
}
\keyword{data}
\keyword{simulate}
